enum for runway surfaces

+++ src/api/resources/airport.py
@@ -2,9 +2,8 @@ from pydantic import BaseModel
 from typing import List
 
 from api.resources.runway import Runway
-
 class Airport(BaseModel):
-    altitude: float
+    altitude: int
     runways: List[Runway]
     total_arrivals: int
     total_departures: int
